% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event_tracking.R
\name{event_tracking}
\alias{event_tracking}
\title{Event Tracking}
\usage{
event_tracking(
  rev_product_ids,
  rev_session_id,
  rev_username,
  event_type,
  product_properties_df,
  desired_properties,
  installed_start_date,
  installed_end_date
)
}
\arguments{
\item{rev_product_ids}{A vector of Revenera product id's for which
you want active user data.}

\item{rev_session_id}{Session ID established by the connection to
Revenera API. This can be obtained with revenera_auth().}

\item{rev_username}{Revenera username.}

\item{event_type}{Choose type of event tracking, one of "lifetime", 
"basic", or "advanced".}

\item{product_properties_df}{Data frame with available properties 
for all product ids. Can obtain with the get_product_properties function.}

\item{desired_properties}{The property names of the metadata you want
to collect.}

\item{installed_start_date}{Date object for the starting date of 
product installations.}

\item{installed_end_date}{Date object for the ending date of 
product installations.}
}
\value{
Data frame with selected properties for each Client Id.
}
\description{
Get aggregated data across clients on different types of event usage 
over time for all product ids in a list.
}
\details{
Revenera provides separate endpoints for lifetime event tracking, 
basic event tracking, and advanced event tracking. More information
on each of these is available at 
<https://docs.revenera.com/ui560/report/Content/helplibrary/Event_Tracking_Reports.htm>.

Not all possible parameters for each endpoint have been included in
this function. For now the function allows specifying the products,
date range, events (if applicable), and global filters (if applicable).
You can specify a start and end date but Revenera does not store
an indefinite period of historical data. In my experience this is 
three years but I do not know if this varies on a product or client
level.

It is not recommended that your username be stored directly in your
code. There are various methods and packages available that are more 
secure; this package does not require you to use any one in particular.

WHAT IS THE API LIMIT??? For the same reason you are encouraged to break your request into
smaller chunks using the install dates and/or splitting up your
product Ids.
}
\examples{
\dontrun{
rev_user <- "my_username"
rev_pwd <- "super_secret"
product_ids_list <- c("123", "456", "789")
session_id <- revenera_auth(rev_user, rev_pwd)  
product_properties <- get_product_properties(product_ids_list, session_id, rev_user)
sink("output_filename.txt") 
sink(stdout(), type = "message")
client_metadata <- get_client_metadata(product_ids_list, session_id, rev_user,
product_properties, c("Property1", "Property2"), start_date, end_date)
sink()
}
}
