% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logout.R
\name{logout}
\alias{logout}
\title{Remove authorization cookies (logout).}
\usage{
logout(rev_username, rev_password)
}
\arguments{
\item{rev_username}{Revenera username.}

\item{rev_password}{Revenera password.}
}
\value{
Nothing (successful logout), or an error message.
}
\description{
Use this to remove the current authorization before re-authorizing.
}
\details{
It is not recommended that these values be stored directly
in your code. There are various methods and packages
available that are more secure; this package does not require
you to use any one in particular.
}
\examples{
\dontrun{
rev_user <- "my_username"
rev_pwd <- "super_secret"
revenera_auth(rev_user, rev_pwd)
logout(rev_user, rev_pwd)
}

}
